requirements.txt
Django==2.2.1
django-environ==0.4.5
pytz==2019.1
sqlparse==0.3.0
Django==2.2.1
django-environ==0.4.5
djangorestframework==3.9.4
pytz==2019.1
sqlparse==0.3.0
